requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-huggingface
pypdf
chromadb
python-dotenv
sentence-transformers
langchain-text-splitters
unstructured-inference
unstructured-pytesseract
pikepdf
unstructured-inference
unstructured-pytesseract
pikepdf
streamlit
langchain
langchain-groq
langchain-community
langchain-huggingface
langchain-text-splitters
chromadb
pypdf
sentence-transformers
python-dotenv